Override the Faraday Client to prevent the body on DELETE
The default API client seems to like sending an empty object
in the request body, which Square doesn't like. In a previous fix
the parameters were exposed such that they could simply be set
to `nil` - I'm unsure the best way to accomplish this, with the new
APIMatic setup since I last encountered this issue. This simply prevents
the empty object from being set on the body, and thus the body is truly
empty for the server to handle as JSON.